live-preview: s/unwrap()/unwrap_or_default()/ in eval.rs
I did nto bother so far as we have code we know compiles and can thus
not trigger most of the unwrap calls, but then we do have places
where we have less information than the compiler and that might trigger
an unwrap.

So let's be safe and just `unwrap_or_default()` (which is typically
return `Value::Void` here). That is fine as we need to expect
`Value::Void` a lotmore often than in the compiler anyway due to
not being able to evaluate all the things.