live-preview: Update palettes when selection changes
Selection might change to something in another file,
which will have other imports. So refresh the palettes
on selection changes.

We do not (yet?) want to support all possible globals,
auto-importing the needed types. We are afraid that this
will make a huge mess in no time when users end up having
slint files open that belong to different UIs!