live-preview: Add warning about missing panics doc to eval.rs
That is supposed to help catch some cases, where functions panic,
but it missed the two `expect(...)` calls, so it is definitely not
a complete solution.

I'd love to fuzz this, but AFAICT `rowan` does not support the
`arbitrary` crate, so I can not have the fuzzer create random
Expressions as input.

Fix the few clippy warnings in the LSP.