live-preview: Have a handler for `debug(...)` in slint code
Allow to have a handler for `debug(...)` in slint code when running
in the interpreter.

When the live-preview is active, print the debug output prefixed with "DEBUG>"
and show the debug messages in the UI.

This makes it unnecessary to ask people to look at the console in Slintpad.

Start work on a ConsolePanel to show the log data in the UI