live-preview: lazily compute the palette
Computing the palette can be quite expensive if there are many globals
with many big struct (in the case for example of a big figma export).
This is done for every refresh or selection making the UI not responsive
at all.

This change lazily compute the palette when needed by the color picker.