test-data now creates ALL data upfront
This is a new approach that first creates the main data. It then scans this data for variable references that point to deleted variables and/or variable collections. Figma then lets you get those collections and variables via another api and we then insert them back into the data.

This means the converter can be simplified in a later refactor as there was complexity in the logic to handle this problems after it was given the data. Now the data is complete the convertor can be simpler.