Fix UB when a change tracker gets deleted as a result of that change tracker evaluation
Dont drop the ChangeTracker's BindingHolder if it is currently being
evaluated as we still need it later.

There was also a bug in DependencyListHead::clear as it was not properly
updating all the field, and the use of for_each was problematic if nodes
gets delete as we iterate.

Fixes #8741