material: Stop favicons being hashed (https://github.com/slint-ui/material-components/issues/60)
Favicons are not showing up in the google search results for docs and this material site. It's possible this is because the favicons are being converted into a file name with a hash e.g. favicon_233242.ico. This ensure they are in the public folder where Astro won't hash them and change the name. This might fix the google search results issue. 

Imported from https://github.com/slint-ui/material-components/commit/385065e768e330a7dac4db34f7abfbe70e94402c