LineEdit: fix height for width dependency
There is a binding loop that cause a failure in the property editor of
the live preview.
Previous version of the LineEdit didn't have a dependency between the
height and the width.
Apply an explicit width binding on the recently added image to prevent
that.

For the fluent style, also change the size of the `x` to something
closer to the actual fluent style (picked the value from fluent's figma
design)