Drag&Drop: don't keep all the touch area that cross the mouse with has-hover
Before this patch, if you drag from somewhere and go over some
TouchAreas, they all would be marked as has-hover and they like that
untill we go over them again without a drag.

There is still an issue that if we release the drag over a TouchArea, it
won't be marked as hovered.