Skia: Fix regression in text size calculation
When computing the implicit height of a Text, we call text_size() on the renderer with the previously computed width, which in turns creates a text layout and *should* wrap text according to the width and return the height.

Before #8138, text layouts created with NoWrap were still wrapped and produce a height as if it were wrapped.

After #8138, NoWrap would return a height as if the text were on a single line. Unfortunately, the text_size() call used TextWrap::default() as wrap mode, which is NoWrap. That was previously an accidental no-op, but now isn't. So we need to pass TextWrap::WordWrap to get a reasonable height.

Fixes #9053