winit: Avoid panics due to zero width/height windows
Don't pass zero min/max sizes to winit, as from there on it might get
back to us, and we'll try to create zero sized rendering surfaces. But
even if we bailed out then, it would be unusable for the user.

For example, this avoid panics or crashes in the live-preview when
editing window/layout code that ends up with a zero min/max height.

Fixes #8926