unit_product_length_conversion: early return for ([], [])
This is a very common case. Any conversion from float to int will
trigger this function with empty vecs, for instance.
When building a tiny example app, this is called 304 times with
empty vecs (while parsing all the builtin .slint files)

The main motivation for this is readability / debugging more
than performance though.