CI: Streamline and fix windows tool binary artifact creation
Instead of wrongly fixing up the .zip archive name in the prepare_release phase (and missing a space as per #9050), let the slint_tool_binary workflow create the correct .zip archive, then let upload-artifact archive the zip (instead of the directory), so that unpacking on the release job side will retain the zip file.

This way it's handled the same as for .tar.gz (where download-artifact doesn't unarchive) and the file name is decided in one place (slint_tool_binary).

Fixes #9050