esp-idf: Permit usage of slint::invoke_from_event_loop() before running the event loop
With EspPlatform, we need to initialize the FreeRTOS task used for
posting events before calling invoke_from_event_loop(). This is now done
basically in `slint_esp_init()` (well, the platform ctor).

The assumption preserved here is that the task that sets the Slint
platform is also the same task that ends up spinning the event loop.

This should fix the following (pseudo-code):

```cpp
void main()
{
    slint_esp_init(...);
    ...
    // used to crash because EspPlatform::task is only initialized in run_event_loop()
    slint::invoke_from_event_loop(some_functor);
    ...
    slint::run_event_loop();
}
```