CI: Fix cargo build -p slint --features renderer-skia,renderer-femtovg-wgpu build
Enabling the wgpu feature in i-slint-renderer-femtovg implicitly enables
the i-slint-core/unstable-wgpu-26 feature, which means the ImageInner
enum has a WGPUTexture variant. That one in turn must be matched in the
Skia renderer, when looking at the ImageInner variants. Fix this by also
enabling the unstable-wgpu-26 feature in i-slint-backend-winit and thus
i-slint-renderer-skia. This doesn't automatically enable the API though.