Fix regression in the Path element when used with inheritance (#9176)
After commit ef2e6491cddd619fa00af321d379b77071635b0a we end up processing Paths multiple times, and thus we might end up attempting to replace an elements binding. Don't do that if no new elements are declared, but yield an error if attempting to do so. We could support this in the future (appending), but for now it's better to produce an error.

Fixes #9170