LineEdit: Fix show password icon being out of sync with fluent
The problem was that the `LineEditPasswordIcon` had an internal property
that would get out of sync when the component is re-created.
The component does get re-created in fluent depepending on the focus,
and this is the case especially with wasm because of the hidden input
that takes the focus.

Fixes #9225