Windows: Use `WM_SETREDRAW` to disable redraw when rebuilding `muda` menus
When the state of menus are changed, the `muda` representation of menus is completely torn down and rebuilt.  On Windows, this can create a slight flicker unless the act of redrawing the window is disabled.  With this change, we use `WM_SETREDRAW` to disable redrawing while the menus are being reconstructed.

It isn't clear to me if other platforms have a similar problem.