FemtoVG: Fix panic when used with WebGL
In earlier refactorings, passing through the id of the HTML canvas element got "lost". Restore this by handling it in the place where we also have access to the element.

This restores rendering as part #9345 , but the size of the canvas is still wrong. But if resized manually you can see rendering now :)

cc #9345