Fix incorrect size of several wasm demos
Commit 1c73144b09d263ebe20cf5e7cf500e7d4d5ac3b5 (famous resume protocol) change the overal timing and sequence of window creation. As it happens now, when
we set_visibility() is called and we try to preserve a "browser" assigned canvas size,
we do so even for windows where we decided we want to go with the preferred size. This restores the todo/underlay/etc. demos.

With commit bcfa1eaa2ea9f90c7bf333780e8467053f3a3c17 we now end up computing current_size as (10, 10) for the slide puzzle (which has no min-width in Slint),
which differs from (0, 0). But we really shouldn't try to assign a size in the first place.