vscode: fix the crash reporter
The rust process will use several `writeln!` on the file to write out
the panic information. This may take some time and vscode could be
reading these files before. So add a small timer in vscode to make sure
we do not try to send the telemetry before we have all the info