lsp: fix removing binding of declared property
When we have a property such as
 `property <int> foo: 45;`
And the live-preview user wants to reset the value of that property,
we should keep the property there.
So the result should be
 `property <int> foo;`
and not remove the property completely as the old code was doing,
otherwise it is likely that this wouldn't compile if the property was
used anyway.