text item: improve partial rendering when unwrapped text changes
When Text elements in a vertical layout change their text, their height would be marked as dirty and thus the entire layout is invalidated. For text elements that don't wrap, the height solely depends on the font metrics.

Note: The text element itself might still *render* above and below that.
The dependency tracking here is covered by the bounding_rect() function
and the individual text item's property dependency tracker in the
partial renderer.