MacOS 26 app icon for Live-Preview (#9442)
MacOS 26 introduced Liquid Glass and a change to app icons. They now require a background, and to support 'default', 'dark' and 'transparent' variants. Apple Icon Composer was used to create a new icon, then exported to Xcode which builds the file into a *.icns file. This shows a default icon on systems older OS's and a correct looking and behaving icon on the latest OS.