GridLayout: keep row/col/rowspan/colspan properties specified by the user
When an item in a grid layout specifies row or col, this property is no
longer removed by the compiler.

That's the one liner that changes removes() with get(), but the side
effect of that is that the code that checks that no row/col/etc.
property is still there afterwards (to give compile-time errors)
had to be improved to check if this is an item inside of a grid layout,
since the property stays now.