Warn about and remove statements without effect
In code-blocks, only the final expressions value is actually used. All
other expressions must have side effects or are otherwise useless.

This usually indicates an error by the programmer, so emit a warning to
notify the programmer about the issue.

Closes #4954

Example from the associated issue:

This Slint code

```slint
FocusScope {
    key-pressed(event) => {
        if (event.text == "a") {
            accept
        }
        reject
    }
}
```

Looks like it would return accept, but in fact, the statement has no effect.

Changelog: The slint compiler now emits a warning if a statement is without effect