Docs fix: the default value for row and col in a GridLayout isn't 0
If you set row and col to 0 for multiple cells, they all end up at 0,0.
So the default value isn't that. It's more something like "auto" where
the col is one more than the col of the previous cell, or 0 if there was
a Row in between.

But it's not like people can write col: auto, so I hope the made
up "auto" keyword here isn't too confusing; the alternative would be
to remove the printing of a default value from the docs, not sure if
that's possible.