CI: Explicitly install the stable toolchain, ignore mise
This should fix the CI in that it will install the full stable
rust toolchain.

Right now we have no current stable toolchain on the machine and rely on
cargo implicitly installing it in the minimal profile when running
`cargo xtask`.

We can not rely on mise to do the right thing here: Apparently it
finds rustup installed and considers the rust configuration to be met,
independent of whether the stable toolchain is up to date or not..