fontique: Fix text rendering with wasm32-unknown-unknown and QNX
When targeting wasm32-unknown-unknown, there's on system font fallback
on fontique, because there isn't a way to discover and fetch fonts from.
With QNX there are ways of discovering fonts on the file system, but
neither we nor fontique do that yet.

So do what we did with fontdb: Include a trimmed version of DejaVu Sans
and install it as fallback for everything.