parley: Apply elision only on the last line
When breaking multiples lines of text and the last line can't be broken anymore but exceeds the available width *and* overflow handling is set to elide, *then* apply the elision glyph substitution.

Also, centralize more of this inside just one glyph iterator, so that
re-use in the other renderers will be trivial. This also avoids one
extra vec allocation when collecting the glyphs.