Material: Stop the APK building for external PR's (#9607)
Currently PR's that trigger the Material CI to run will fail for 
external PR's due to how the APK signing key needs to be secured. This 
PR skips the APK build for external PR's as we build it in a manual run 
only when we publish a new Material update.