Qt: Emit scrolled callback when dragging scrollbar
This is supposed to be emitted whenever the user performs an
intentional interaction, which works whenever you scroll the Slint
Flickable with the scrollwheel. Dragging the scrollbar on the Qt
style didn't do anything, since that was never hooked up to this
callback.

I only implemented this in the Qt style, because that's the only
one that uses the NativeScrollView element.

See #9574.