Some small ScrollView documentation changes (#9591)
* Use standard defaultValue documentation for ScrollView scrollbar policy

Instead of writing it as a literal documentation comment, we can specify
defaultValue on the SlintProperty itself like other items do.

* Improve documented limitation with ScrollView and for loops

It's now a separate note block, so it doesn't blend in with the rest of
the paragraph. I also turned the issue number into a real link, like we
do elsewhere already.