Compiler: fix joining relative path to `""`
When using the `slint!` macro, the path given by the `Span::local_file`
are relative to the workspace.
When there is only one crate in the workspace, the path to Cargo.toml
will be just `"Cargo.toml"` and its parent will be the empty path.

Fixes https://github.com/slint-ui/slint/issues/9653