Add new widget test for ScrollView, test the scrolled callback
As brought up in 7d038eb9e293b243880269e65f6799bb69998d83, the bug
I was fixing should really have a test for it - so I added one.
Testing the scrolled callback is a bit odd to handle because not
only does each style have a different scroll handle but it's a bit
dubious when/how many scrolled callback emissions there should be.

I was able to find a sweet spot for both of these problems, and
although the test case is a bit vague it does catch when a style
doesn't emit the callback.