Link directly to individual global enums/structs in documentation (#9600)
Previously, we sent users to the top of the page whenever they
click on a property that refers to a global enum/struct in our
documentation. However the page is already filled with permalinks
we can use, so I made it so we assume there's a permalink.

If the permalink doesn't exist, it would just send them back to
the top of the page anyway which was the previous status quo.