Remove Text's bounding rect implementation
Since every renderer is now always clipping the text, we don't need to
account for out of bound drawing.

The other optimization of trying not to draw things if the bounding box
is smaller is not really a good one since computing the bounding box is
itself too expensive and not cached.