build: use absolute library and include paths
If CompilerConfiguration is configured with relative library or
include paths assume that they are relative to the manifest directory.

Update the configuration to contain only absolute paths so users of
the live-preview feature who configured the compiler with relative
paths can run their application from directories other than the
manifest directory.