Fix scrolling for ListViews with varying item heights
The viewport_y of a ScrollView is saved when a mousebutton is pressed.
(inner.pressed_viewport_pos in flickable.rs and pressed-value in the
scrollview.slint files).

The saved value is used to compute the new position when the mouse is moved.

When a ListView loads new items and the average height of the items changes,
the viewport_height and viewport_y of the ListView change and
the saved value has to be adjusted to avoid 'jumping' during scrolling.

TODO:
- The fix is not implemted for the qt style yet.
- Maybe it would be better to check for property changes
  instead of saving the viewport size?