fontique: Fix panic when using Text.font-metrics in wasm builds
The lack of generic families being defined means `query_fontique()` returns None, which we should gracefully handle.

In the case of #9700 that means - with this patch - that the gallery doesn't panic anymore, but the icon in menu items aren't visible due to a zero height, due to zero valued font metrics.