fontique: Fix font metrics queries with wasm builds
In wasm builds, we're responsible for populating the empty fontique collection with the one font we have. So register it with matching generic families.

Also, when creating the fontique query for a FontRequest, align the families in the same way as we do for text layout when specifying the families to parley's FontStack style.

This brings back the icons in menus in #9700 for wasm, as now the metrics are non-zero and thus the icon gets a height.