Python: Fix interaction with garbage collector when models are held in structs and properties
Models in Python have a strong reference to the Rust type, which have a strong reference back to the Python model. Therefore the __traverse__ and __clear__ protocol is implemented, to handle this cyclic reference.

The same protocol needs to be implemented for structs itself as well as basically all properties held by the interpreter.

Fixes #9706