parley: Fix line culling when word-wrap is enabled
With the following test case

```
export component TestCase inherits Window {
    Text {
        wrap: word-wrap;
        text: "Hello World";
        overflow: elide;
    }
}
```

The Text would get a max height at rendering time, for me it would be 12
pixels. The max_coord of the line metrics is 12 as well, but we'd cull
the line - erroneously. This patch changes it to use >= because the
max_coord is not an actual coord but a height (it's baseline + descent),
and it applies the culling only when we don't do pixel clipping.