software renderer: Fix severe slowness when using system fonts
With fontique collections, we rely on the shared source cache to be effective, to avoid repeatedly loading new fonts when creating new fontique::Collection instances. That cache however stores weak references to the blobs.

For parley based rendering, we keep one single fontique collection in the thread-local CONTEXTS in sharedparley.rs, which basically keeps strong references to font blobs we use.

When not using parley, we discarded the strong reference to the font blog after creating the fontdue font, which unfortunately meant that the font was discarded from the shared cache and loaded *every* time we queried the collection.