software renderer: Remove rustybuzz dependency
We're not going to go down the route and do full text layout with our own text layouting, so for the system-fonts case without parley, eliminate the rustybuzz dependency.

It'll remain a dependency with resvg, but that might go away in the future. The skrifa/read-fonts dependencies aren't new, they're part of the parley dependencies.