partial_renderer: Fix `visible:true-expr` item being always dirty
The issue is that an element with visibility creates a `Clip` item as a
parent with a size of 0x0, and the clip property.
When the visible property becomes true, the tracker for this Clip item
will be dirty and will force everything inside to be re-drawn.
But since the `Clip` itself is no longer clipping and its size is 0, the
special code in `iterm_rendering::render_item_children` that handle the
clip specially will not kick in, and the `Clip` itself will not go
through the partial renderer, so its tracker will stay dirty forever