Improve Rust caching and setup in autofix.ci
The rust tooling was being installed twice in auto fix.ci. First by MISE
and then again in a dedicated step. This now leaves the Rust install to 
MISE.

There was no cache being used at in in the autofix step. This fixes 
that. If at any point the cache needs to be reset increment the -v1 part
of the cache key name.