Scroll Into View for Focused Item
Implement a "scroll to make points visible" function to Flickable, automatically scroll to the bounding box of an element when it gains focus.

ChangeLog: When resizing the window, try to keep the focus item visible if it's in a Flickable.
ChangeLog: When focusing an element that's not fully visible, try to scroll a parent Flickable so that it is.

Fixes #3067