system-testing: Do not panic on TCP communication errors
On any TCP communication errors between AUT and test server, do not
panic with unwrap() anymore but just close the connection and print
the error to stderr. So the application will not panic anymore when
the test server quits.

Reason is that the panic randomly happened during normal test execution,
probably during test shutdown phase. Even though the tests were
successful, on macOS a crash reporter dialog was raised due to the
panic, which is annoying. Fixing the panic will prevent the crash
reporter from unintentionally showing up.