autofix: Stop type-checking the vscode project (#9891)
Typechecking the vscode project needs a wasm build of slint so the types
are generated. This stop that so autofix should now run in ~1 minute.

If this works the next stage will be to run autofix before all other
jobs and only kick them off once its successful. The reason for this is
we only have 20 runners. Autofix and the normal ci.yaml jobs start. This
takes up 10 to 15 runners. After 5-6 mins autofix often finds a problem.
Makes a new commit and now the other ~15 jobs need to be cancelled. So
in theory we can reduce pointless jobs running.