Fix concurrency issue with autofix and ci (#9901)
A previous PR removed the concurrency group from autofix.ci as it 
conflicted with the one in ci.yaml. 

A side effect of this is that a new PR won't cancel a previous versions 
run until the autofix part had completed. This only takes a couple of 
mins, but it's still keeping many runners busy on a task that should be 
cancelled.