Tweak how the Rust cache works in CI (#9902)
This adds an option so any job can opt out of using the cache. It defaults to using the cache so right now everything that was using the cache will continue to do so.

The second change is that as we are well over the 10GB cache limit. So we will no longer save the cache created in a PR branch. As only the master branch cache is shared this PR branch cache could only benefit subsequent runs of the same PR. If this turns out to slow PR's down it can be easily turned off later.