Clean up more of the docs cache (#9910)
No need to cache the docs run. It was creating a 2GB cache file which uses up 20% of the 10GB allotment we have. It could be reduced to 1GB, how it's not really needed as without a cache this job only takes around 20 mins and could be split up to complete even faster.