Fix Frontboard issue in nightly tests (#9962)
FrontBoard is a system that detects what app are on an iOS device or 
simulator so they can then be launched. It can be slow to detect a newly
installed app leading to a race condition in CI where the action tries 
to launch a successfully installed app before it's been registered. It 
fails to launch and the CI fails with it. Now loop for a few seconds 
retrying to launch before failing after 30 seconds.