Try and resolve issue with multiple simulators (#9965)
This PR does 3 things.
1. The iOS simulator which is only needed after the build was booted before the build. 
It is so resource heavy it causes the build to take 20-30 minutes. The simulator is now
booted after the build and the build itself now only takes 5 minutes.
2. There are multiple versions of the same named simulator for the multiple versions of
Xcode. It seemed somehow the wrong simulator could be selected for booting/installing/etc.
We now get the UUID of simulator and use that to launch it. That seems to have made the 
job work. But as it fails randomly we don't know for sure yet.
3. The simulator would then be killed by the clean up process as the end of the run and it took
~40 seconds. There is now a step that tells the simulator to shutdown and that only takes 
~7 seconds.