CI: Attempt to simplify docs generation
In each C++/Rust/etc. docs step, attempt to generate an artifact that
has the directory structure suitable for the website right away, so
docs/$lang.

That way, eliminate the combine-deploy step and move the 404.html and
index.html generation to the publish_artifacts step, eliminating the
need to clone the website repo and running the generator for each build
in the master branch.